+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v1.27.0] - 2026-04-28
 ### Added
 - Navigation menu on the header capabilities with `site.navigation`
+- Optional hidden mobile menu label support with `site.menuHideLabelVisibility`
 
 ### Changed
 - Fixed dev vulnerabilities